Add ansible variable detection for plays
The current variable detection only apply once at the begging
of the run script. At that time, the user variable file might not
exist yet.

Without this patch, the user_variables could not be loaded on the
first deploy.

This is a problem, as this would break installations.

This fixes the problem by ensuring that the user_variable files
are detected and passed at step which requires them.