Use the same way to run scripts
The scripts were only using overrides if we needed to.

This is a problem, as deployers could not override software
we didn't plan to override.

This should fix it by using a standard way to override things,
and provide default empty overrides, that can be used by
deployers.

This also removes the glance-install.yml task file, as it is not
used anymore.