Provide default overrides
By default, the role allow the users to pass their overrides,
making the system very flexible.

However, if no overrides are given the role tasks would fail
on undefined variables.

This could be fixed by adding |default({}) everywhere the
variables are passed, but this will then lack documentation
in the future.

Therefore this patch fixes the undefined variables by
defining them in the defaults/ allowing us to use defaults/
as a mean to document the features of this tooling.