Do not assume deployer is localhost
Without this patch, deployment would fail on a non SLE (for
example a macbook) localhost.

This is a problem, as we confused two roles in this tooling, the
developer/ci node, and the OSH deployer, focused on that purpose.
They can be the same node, or be different.

This fixes it, and clarifies the usage in the documentation.