Allow simpler user level overrides
Without this path, the default overrides provided contains
the common image overrides.

This is a problem, as it means the deployers who want to override
a single image will have to provide the entire list of overrides.

This fixes the problem by ensuring all the templates have all
the content at all time, so users can override the way they want.

To avoid not repeating ourselves, this patch uses a single
common image overrides files, but it becomes very complex to
follow.

A more surgical approach of the images overrides should be taken
in a later patchset.