Allow externally provided ceph mons
Under certain cases, like running on a cloud with floating ips,
the generated ses-config only contains the private IP, not the
IP it should be used to reach the final node (if different
networks or port security for example).

If we are only using the ceph mons from the ses-config, the
cloud deployment would then fail, because the ceph mons are
unreachable.

This is a problem but very situational.

This patches fixes it by allowing a deployer to override the
default provided ceph mons, by providing its own list of
ceph mons, each element would be composed of ip:port.