Draft documentation for semi-automated install
Here's a draft of how to use Stages 2 & 7 on a manually installed
SES and CAASP. This is a little incomplete in the "how to install
CAASP/SES" area right now - I'm of the opinion that we can either
build on it or obsolete it via KVM with kubic-automation.

Note that this install process is not currently guaranteed to produce
a full working environment due to disagreements re the inventory-ses.ini
format between Stage 2 and Stage 7 the last time I worked with it
(which presumably also break run.sh deployment on engcloud :() But it
should give everyone an idea of what needs to be done and can be refined
as we go.