Remove rbd application enable in Stage 2
Enabling the rbd application on the pools here causes the
helm service bringup to fail since an application with a different
name has already been enabled for that pool. Remove this for now to
get it working; look more in-depth at pool creation, etc later but
leave it as is for now since it's not a blocker.