Building suse image for openstack services
This change is using openstack-helm-images and loci upstream patches
to build openstack services images. So added support for pulling-in
those patches. If a given patch is merged in master, then that change
is not checked out as its already available via clone of that repo with
master branch.

Currently its building stable/queens images.