Simplify dev patcher
The suse-dev-patcher role was filled with more complex data structures,
including repetitions of current structures. It had also unnecessary includes.

This is a problem, as its harder to maintain over time.

Instead, this moves as many tasks as possible directly in the main.yml, and
keeps the include for clarity when deploying (using a task title that can be
looped over for example). It also ensures ONE data structure for the
upstream repositories is set in place for later re-use.