Refactor run.sh
Run.sh was containing a series of functions that requires
adapting should one backend deployment (openstack, baremetal, kvm) change.
It was also hard to know where to add steps, should steps be added.

This was a problem, as we needed to add a step in CI to add VIP handling.

This fixes it by making run.sh into a caller of different scripts by
convention, based on environment variables (by default things run on openstack).
This allows a cleaner separation of multiple steps,
and run.sh wouldn't require code change for different base deployment
technologies, like kvm or bare metal.

This resulting work was then leveraged for CI to add a VIP management step,
and load it into the extra variables. The extra variables used in the past
were not compatible with the ansible-runner standard, so this patch also
cleans the ansible scripts.

To have a VIP in the CaaSP cluster on openstack, multiple steps
had to be taken:
1) Disabling the port security on caasp nodes, so that a vip
   can be assigned into any caasp node, and the traffic would not be
   blocked by engcloud's neutron.
2) Creating a VIP to "dedicate" an IP for the VIP, should we
   want to run multiple jobs at the same time.