Do no skip the include step
If the include step is skipped, the dev patcher will fail, as
the variable on which to apply the dict2items filter is undefined.

Even if the task is skipped, the filter would throw a Jinja exception
causing the play to fail.

This fixes it by ensuring the include always run (even if tag
filtering on the play is applied), and by ensuring the dict is
empty by default.