Fix the rbd user variable
When ses_cluster_configuration's content was adapted, the user of
a pool was renamed from rbd_store_user to rbd_store_pool_user.

This replaces all the references of the former to the latter.
At the same time, template was incorrectly using an attribute
notation instead of dict lookup, this fixes it at the same time.