Update OSH to latest
The December 4th version of the code relies on dumb-init, which
is not present in the latest images. This ensures the latest
image can be used.

The new images required a change in priviledged-cluster-role-binding,
which was therefore updated (else nginx couldn't bind on 0.0.0.0 in the
openstack ingress controller).

This also keeps the code frozen to preventing surprises in new code,
which will increase our velocity.