Bring back code for importing SES configuration
This commit reintroduce the generation of ses_config.yml which is needed
as input to osh_deploy. It's based on the code that was removed as part of
commit 2afb55084 and introduces the new ansible role "ses-osh".

We also move the code create the ceph pools and import the admin keyring
from the main playbook to the new role.