Rename playbooks to be more consistent
This moves the playbooks to be named in a consistent way.
Each playbook is named as <deployment mechanism or generic>-<action>.

This simplifies the readability by making it easier to know what each
playbook is doing. Similarily, each "action" from run.sh is mapped to a
playbook by having it named, for example, for openstack deployment mechanism,
openstack-<action from run.sh>. It's therefore easier to follow.