Rework ssh wait loop and knownhosts update
The addition of the host keys to the knownhosts file frequently failed
for the SES and CaaSP nodes because ssh-keyscan didn't return any valid
keys for the nodes yet.

This commit changes the wait condition to wait until sshd is answering
on port 22 and additionally adds a retry to the "ssh-keyscan" call for
running the command until it actually returns something. (Unfortunately
"ssh-keyscan" always returns a 0 exit code it seems)