Split caasp node enrollment
The process was to enroll caasp nodes and immediately modify
the caasp workers to allow openstack deployments.
However, when developer mode is enabled, the current process would enroll
the caasp nodes completely, then only run the dev patcher/build images, and
not touch the caasp worker nodes again.

On top of that, merging the two plays break the expectations for BYO CaaSP/SES
story: The users who brought their own CaaSP/SES clusters still need to
run that generic-setup_caasp_workers_for_openstack.yml playbook.

This splits the action in two, one which can run at the setup_hosts step
(equivalent of the bring your own), and one for the rest of the steps
(setup_openstack). The deploy_osh is kept as standalone. The run.sh actions
were sorted in a "chronological" order.

In that first action of the setup_openstack, we'll be then able to generate the
certificates for a BYO cluster + developer case, to be then consumed
by the developer plays (patch upstream/build images).

In CI, to avoid having flapping certificates, we could run the playbook
to generate the certificates without running the rest of the play. That
would be the equivalent of "BYO certs" too.