Use built registry
This points the worker nodes to the developer's built registry.
To use it, /etc/hosts needs to be changed, and the certificate
of the registry needs to be copied over.

At the same time, the build_registry variable was incorrectly defined by
being in the vars/ folder. The var's value was a live evaluated one based
on facts, which is generally a bad idea to have, because it depends on the
scope of the play.

This therefore moves the build_registry to a live evaluated variable in each
play. It also have shown that some of those previously defined build_registry
was incorrect (having ports where no ports were required for example).