Re-use setup-caasp-worker task when possible
The generic-build_images step configures a secure docker registry.
Therefore the registry consumes the given cert/key already.

However, docker push upload will fail, as the host itself doesn't know about
the certificate. We have to then apply the same logic as the caasp workers
for the osh-deploy node where the docker push is done.

For that we re-use the role setup-caasp-workers. However, a refactor was
necessary to prevent loading unnecessary tasks.