Add the cert generation in the process
Assuming developer mode is enabled, the cert generation would never be
triggered, as it's not called in the plays.

This changes it, as it includes, as first step after the cluster enrollment,
the certificate generation.

The certficate can then be used, in later plays, for the caasp workers, or
for the osh-deployer node.