Applying conflicting changes from ucp branch to master
Syncing common variable change from ucp branch into master.
Now variables are named to socok8s instead of having osh_ or
airship_ namespace in prefix of their names.
Variables which are not shared between and are unique to osh
or airship deployment are left untouched.

Also getting changes from roles which are common and are changed in
ucp branch to bring attention in this review.

Adding group_vars file for common variables shared across various
roles so that they don't need to be explictly included in various
roles where needed. Currently common_vars are used for this purpose
which can be phased out if group_vars work for our need.