Fix CaaSP stack delete
Altough the `Get a floating ip for the workers` task is skipped it
requires to template the loop parameter which fails as one of the varaibles
is defined by a task that is skipped and does not have `stdout`.

This change adds a default value (empty json list) to cover that case.