Simplify templating
Config template has a bug in multiline, and we don't know when
it will get solved.

We can skip config_template and rely on kubernetes has merge
behaviour. For that, we just need to pass the files in the cli,
on the right of the rest.

This adds this feature, and at the same times, cleans the hacks
we had around to use /bin/true to generate templates.