Handle no new revision error to make playbook idempotent
Create/commit config docs action in shipyard fails with error when
there is no change in already existing collection (site design).

To handle that situation, we are ignoring that specific error during
create and commit actions. At this stage, we don't have ways of doing
diff between new update and existing collection docs that's why this
approach is needed.