Split inventory into multiple files of a single folder
For CI (run steps in parallel) it will be useful to split the
inventory file into multiple files, so that we don't have
race conditions during editions.

The documentation and examples is NOT updated, on purpose.
We will keep the "hosts.yml" as a example skeleton file of inventory.
It is easier to include the inventory in documentation that way,
and therefore explain to the users the expected groups.