Work with empty folders
The current run-ansible assumes if a folder exists, the inventory/env
folder would exist with their full valid content.

For CI and for users, it would be better to not have that expectation:
A user might have created an inventory with empty content.
We solve this problem by assuming the folder can exist, as empty, and
we'll create the appropriate folders (env, inventory).