Consolidate ANSIBLE_RUNNER_DIR setup
There is logic in both install_ansible and run_ansible to check if
ANSIBLE_RUNNER_DIR is set and to some create initial subdirectories.
Since install_ansible will be run first, that logic will set the
variable and create the directory, preventing the logic in run_ansible
from ever running.

Consolidate code from run_ansible into install_ansible so that we have a
single code location for setting a default ANSIBLE_RUNNER_DIR value and
creating the directory if needed.