Use nova image for novncproxy_assets
Upstream uses kolla/ubuntu-source-nova-novncproxy:ocata which as best as
I can tell is an image with openstack-nova-novncproxy which in turn is
just a packaged version of the nova cli entrypoint nova-novncproxy.

Thus the standard openstackhelm/nova image should be sufficient. Using
that image also reduces the number of images required to be pulled in.

This deploys fine, but I have not tested the actual novnc in horizon.

[0] https://github.com/openstack/kolla/blob/master/docker/nova/nova-novncproxy/Dockerfile.j2