caasp: add extra configs for apparmor profiles
Some of our containers may need extra configs for apparmor.
For example, neutron-dhcp launches dnsmasq, which needs access
to the neutron config dir and the /var/lib/neutron/dhcp for leases.

We need a way of extending apparmor profiles with custom directives
for the proper execution of our openstack containers.

This patch allows us to add extra directives for apparmor profiles