Create role for common tasks
osh and ses_aio instance creation are using the same kind of tasks,
so we can use a role to avoid repeating ourselves.
To simplify the role, the volumne name for osh/ses_aio will be
extrapolated from the server name.

This raises the fact that ses-osh also has an assumption of
the inventory_hostname, which was therefore removed in the
ses-osh role, to use a variable instead.