Use unique servernames
Currently servernames are non-unique, due to the fact they rely
on PREFIX, which is not unique (only workspace is currently unique).

This is a problem if you are running more than a single environment,
because all the nodes of a same type (for example deployer/ses) will
share the same name if they are on the same network.

This is solved by making sure "PREFIX" becomes a unique variable,
representing each environment.

Because the term "PREFIX" was already used for convenience in terms
of non unique networking, I removed all the reference of derivation
of network from the PREFIX variable.