Generate servername based on workspace
Currently servernames are static and do not depend on workspace.
This is a problem if you are running more than a single environment,
because all the nodes of a same type (for example deployer/ses) will
share the same name.

This is solved by making sure the nodes have a generated name, in
the form <keypair name>-<random number>-<suffix>. Using the
keypair name allow to easily track the owner of the server, and
suffix makes it easy to know which kind of server it is.

However, generating name still need to keep idempotency: I must be
able to track back the servername if I want to delete the server
later. This is done by ensuring the random number is not really
random, by using the workspace folder as a seed. This keeps
the server name stable for deployers of a single environment,
while allowing name rotation for CI.