Fix copy of pegleg output
The pegleg output is generated on the remote host and should be copied
locally to the host.

The current code works when ansible is run locally on the deployer
node; but if soc-deployer is a remote host, then the pegleg output
generated in the previous task will not be available on the local
host.