Update velum-bootstrap blob to 510a4c668c
This updates the velum-bootstrap tarball to the latest level from the
the release-3.0 branch:

https://github.com/SUSE/caasp-automation/tree/release-3.0
SHA1: 510a4c668c276985218f40b31ad5114b36260541

This should fix issue with the kubeconfig refresh that we have been
experiencing lateley.

It also updates files/tools/generate-environment to work with the latest
changes in the above code (minionID got renamed to minionId).