Added argbash templates and Makefile
This patch adds generic argparsing via the argbash generator
This enables a nice help output and explainations of the inputs
and the evironment vars as well as the available commands to run.

You can view it here:
https://github.com/matejak/argbash

To generate the run.sh script do this:
make
This will pull down argbash, install it, and then generate
the run.sh, which has the new argparsing and help output.

This patch updated run.sh to use the argbash arguments.  You only
need to make run.sh if you need to add new arguments to the script.

By default ./run.sh with no command will run setup_everything

View the help with
./run.sh -h

To view the list of available commands, just give it a bogus command
./run.sh bogus

The _parsing.m4 defines the 3 arguments
command = which command in run.sh to execute (deploy_ses, deploy_caasp, etc)
environment vars
DEPLOYMENT_MECHANISM - if not set, defaults to 'openstack'
USE_ARA - if not set, defalts to False