Added socok8s.sh for args bash-completion and man
This patch adds the new socok8s.sh wrapper script for run.sh

The socok8s.sh script is generated from an argbash template file
(socok8s.m4) and Makefile.  socok8s.sh has a nice help output and
explainations of the inputs and the evironment vars as well as the available
commands to run.  This also adds bash-completion support and man page.

socok8s.sh is an argbash generated script that knows how to parse
arguemnts on the cmdln, output help and then calls a nearly unmodified
run.sh
By default ./socok8s.sh with no command will run 'setup_everything'
socok8s.sh also shows the used environment vars in the help output.

More info about argbash is here:
https://github.com/matejak/argbash

More detail:
This patch adds a new script call socok8s.sh, which is a wrapper
for the run.sh.  socok8s.sh is generated by the new Makefile

This patch also adds a Makefile, which is used to generate the socok8s.sh
script, it's man page and the bash-completion script that you can source
for your shell environment.

The building of the socok8s.sh is via the new socok8s.m4 file.
This is the argbash input file that describes the arguments that the
socok8s.sh script supports, it's help, description and version.  If you want
to add a new argument that the script supports, then you should:
* modify socok8s.m4
* run make all

The socok8s.m4 defines the 3 arguments
* A boolean option (on/off) for quiet mode
* A boolean option (on/off) for preflight only mode
* command = which command to execute (deploy_ses, deploy_caasp, etc)
* environment vars
* DEPLOYMENT_MECHANISM - if not set, defaults to 'openstack'
* SOCOK8S_ENVNAME - Required env var for base name for all resources
* SOCOK8S_DEVELOPER_MODE - Turn on developer mode?
* USE_ARA - if not set, defalts to False
* PREFIX - The prefix name for all objects created in the OpenStack Cloud
* KEYNAME - OpenStack security keypair name you created.
* EXTERNAL_NETWORK - The OpenStack external network name to use
* INTERNAL_NETWORK -
* INTERNAL_SUBNET - The name of the subnet you created in OpenStack
* DELETE_ANYWAY - Defaulted to NO