Add -q quiet option and developer mode env var
This patch adds to the argbash patch to enable --quiet -q quiet mode,
as well as adds the SOCOK8S_DEVELOPER_MODE environment var for help.

This patch also migrates back to run.sh and wraps what with the new
socok8s.sh to enable bash-completion.

This adds bash-completion script as well as man page