Use local storage for SES and deployer nodes
Up to now the SES and the soc-deployer instances were creating cinder
volumes for their addtional disk. This commit switches them to use local
storage. Which should avoid some overhead that running ceph-osds on top
of ceph-volumes introduced. Additionally it makes the
"handle-nodes-on-openstack" role a little simpler, as we don't need to
worry about volumes anymore.

For the SES node we're switching to a flavor that provides an ephemeral
disk and inject user-data that avoids mounting that disk so it can be
used by the SES deployment scripts later on.

For the soc-deployer this isn't much of a change since it wasn't using
the cinder volume anyways.