Add a Jenkinsfile for testing
The Jenkinsfile will be used for PR testing together with the
github-branch-source[1] plugin.
That plugin does periodically scan the SUSE-Cloud organization and
will create new Jenkins jobs whenever a PR (which must be from
somebody with admin or write access and must contain a Jenkinsfile)
against SUSE-Cloud/socok8s is created.

That way we can also test changes to the Jenkinsfile beside changes to
socok8s itself.

Note: This Jenkinsfile is based on the work from others[2]

[1] https://go.cloudbees.com/docs/plugins/github-branch-source/
[2]
https://github.com/SUSE-Cloud/automation/blob/master/jenkins/ci.suse.de/pipelines/cloud-ccp-full-periodic.Jenkinsfile