Improve file names for creation/deletion of caasp and network
As discussed in [1], we agreed on use different naming for the
deployment steps. Instead of using "deploy" and "cleanup" we use now
"create" and "delete" to make it clear what theses playbook files are
doing.

Note: There are still files that contain the "create" and "delete"
tasks in a single playbook. That will be fixed in a follow-up PR.

[1] https://github.com/SUSE-Cloud/socok8s/pull/289