Assign floating IPs to caasp workers via the heat stack
Instead of using ansible to assign floating IPs to the CaaSP worker
nodes, use the CaaSP heat stack to do that.
That has the advantage that the cleanup of the resources is a lot
easier because you don't have to manually unassign floating IPs when
calling "./run.sh teardown".
Note that we currently don't do the mentioned cleanup so the teardown
function is broken.