Fix extravars file creation when only deploying CaaSP
When deploying CaaSP before SES, the env/extravars file was not
created and the "lineinfile" ansible statement failed because the file
was not there.
That is now fixed because using "blockinfile" creates the
file. "blockinfile" is also used in
playbooks/roles/ses-osh/tasks/main.yml which was responsible for
creating env/extravars . Now it doesn't matter which (SES or CaasP) is
deployed first.

Also move the code to write the VIP to env/extravars to the networking
stack creation. The VIP is created there so it makes sense to also set
the VIP from there.