Pinning cloned repos to specific commit for QA testing
As QA team is now on-board for testing, we need to stick to
specific commits to provide them a consistent snapshot of
locally cloned upstream repositories. Airship internally
(via treasuremap) references specific commit-id for various components
(helm toolkit, openstack services, core components).

May be dev-patcher needs to be enhanced to allow patch set
number in a given review to have consistent snapshot.

All of the commit versions are referring to lastest commit in each
repo master branch to be in sync with current socok8s behavior.

We plan to create git tag on socok8s repos later once this change
is merged and we have a working deployment. Working deployment will
have most of openstack services integrated and are using suse images
with necessary overrides on airship side.