Run caasp update directly after stack creation
This makes the enroll_caasp_workers step a separate task in run.sh and
should a low for better parallelization in the ci later on. I.e. we can
run caasp deployment and update in parallel to the deployment of the
soc-deployer node.
Enroll workers needs to run after that, as it requires the soc-deployer
to be ready.