Jenkinsfile.integration: Add deployment parameter
We want to set a default deployment mechanism. Otherwise we do not run
all the needed steps to do a full deployment.
Also rename the parameter value from "helm" to "osh" to make it clear
that we use the "osh" deployment step.

Note: The first parameter from the list is the default, so we use
'airship' now as default.