jenkins: retry SES step
Seems that sometimes we are hitting an issue when deploying ses which
is fixed by retrying.

As the ses deployment is done by the ses team playbooks and we dont
have any control over those, we cannot directly try to indroduce the
retry on the playbook tasks. Instead we can just retry the jenkins
step, which should be idempotent, to check if its fixed.