airship keystone: override bootstrap script (SOC-9409)
As SOC-9409 mentions we cannot create admin scoped tokens
because the admin role is not assigned to admin user on the
default domain.

This patch fixed that by adding the admin role to the admin user
for the default domain on the boostrap.

This is only a temp patch, as mentioned in the override file
this has been sent upstream, but until that its merged we can
carry this.