Adding note for SES 2 disk requirement for AIO setup.
This doc change is added as per SOC-9323 comments in case user
end up using SES in AIO setup. Generally user is supposed to
follow SES own instruction and guidelienes referenced in docs.
But still we have seen re-occurence of this issue in testing as
its more like a tribal knowledge. So claifying this via this
doc snippet.