Make the OpenStack internal network cidr configurable
When creating a network and subnet on OpenStack, the CIDR for the
internal subnet must be configurable. This is needed for deploying
CaaSP 4 which requires in its terraform variables the internal cidr.

Having this now as parameter allows us to set it in our network heat
stack and use the same value in the terraform variables to create the
CaaSP 4 node.

Note: gateway_ip was removed from the heat stack. If gateway_ip is
omitted, neutron will assign the first free IP address within the
subnet to the gateway automatically.