Add network/subnet/router IDs to group_vars/all
This is later needed when using terraform to deploy CaaSP 4.
Currently the terraform's from skuba want to create the resources by
themself. To avoid this, we can tell terraform (via "terraform
import") that the resources are already available. But "terraform
import" expects the ID (not the name) of a resource. That's why we need
to store the IDs.