Add step "deploy_caasp4"
The new step "deploy_caasp4" does a CaaSP 4 deployment with the CI
codee from skuba[1].
This is still not finished but another step in getting a full CaaSP 4
deployment together with socok8s. Some notes:

- The skuba repo is currently a private repo so a symlink to the local
skuaba checkout in submodules/ is needed.
- The skuba CI code uses terraform to handle the nodes. So terraform
needs to be installed on the node that executes the run.sh script.
- This commit does only setup the nodes in the OpenStack cloud which
are later used to deploy CaaSP 4. So this is WIP and not functional.

[1] https://github.com/SUSE/skuba