Create a group_vars/all/ directory in the inventory dir by default
This is useful when variables need to be available for all hosts. If
this is needed, files with variables can be added to that directory.
That's needed eg. when we create a network stack and want to store the
network resource IDs (for network/subnet/router/...) which we then
need for a "terraform import" call for CaaSP 4.

Note: This is different to the current extravars mechanism. extravars
is loaded via the command line (ansible-playbook -e $extravarsfile)
and using "-e" has always win precedence[1].

[1]
https://docs.ansible.com/ansible/latest/user_guide/playbooks_variables.html#variable-precedence-where-should-i-put-a-variable