Allow using custom openSUSE mirror for Leap
By setting the env var "SOCOK8S_OPENSUSE_MIRROR" it's possible to switch
the repos of the Leap 15 deployer node to a different mirror.

This commit changes the CI job to use the local openSUSE Leap mirrors in
the datacenter, which should give a bit of a speed up for the zypper dup
call. (And also help avoid issues when the external mirror
infrastructure is slow)
Additionally the use of delta RPMs is disabled. When connected to a fast
mirror this doesn't give any benefit over just downloading the full
packages.

(cherry picked from commit 8c4315373f7c19c493f4b54339e6e4348b2c3f88)