Build against Leap 15.1 deployer image
This switches to the official JeOS image for openSUSE Leap 15.1.
Unfortunately this one doesn't have python 2.x embedded so
we need to be ultra correct in specifying the ansible python interpreter
everywhere. On the bright side that actually fixes the reality to
match the documentation (documenation didn't say that python 2.x
is required while it was actually required).