Configurable timeout for airship update action
The CI jobs still run into timeouts every once in a while. This PR
incraeases the default timeout to 120 minutes (which might not even
be too unreasonable for larger deployments).

It also turns the timeout into an attribute so it can be override
via the AIRSHIP_DEPLOY_OPENSTACK_TIMEOUT enviroment variable.