Workspace dir needed for ARA config
The directory needed for ARA config is created either when the virtual
environment is created or in the first call of run_ansible. In case you
are not using a virtual environment and you are using ARA that directory
will be missing and the pre-fligh-checks would fail. This patch adds that
directory in such case.