bump osh repo for deploy_osh
All repos are unfrozen when deploying just osh but openstack-helm-infra
This is currently resulting in an issue where the neutron chart from
master requires a change in osh-infra but infra is frozen to an older
version.

This bumps osh-infra to master like the other repos. This config is
only applied when using deploy_osh so it should not affect airship
deployment as that uses a different repo tracking
(site/soc/software/config/versions.yml)