Introduce TERRAFORM_BINARY_PATH env var for custom terraform binary
For CaaSP4, terraform is used to setup the nodes. But the CaaSP4 code
needs terraform 0.11.X but in openSUSE Tumbleweed there is already
terraform 0.12.X which is incompatible (there are big differences
between 0.11 and 0.12).
To be able to continue with an older terraform version, make the used
binary configurable. Steps to use this:

- Download an 0.11 release from
  https://releases.hashicorp.com/terraform/
- Unpack the binary, make it executable and copy it to your ~/bin/
- export TERRAFORM_BINARY_PATH=~/bin/terraform-0.11

Now continue with the usual ./run caasp4 step to deploy CaaSP4.