Allow CaaSP4 deployment locally via libvirt
With setting DEPLOYMENT_MECHANISM=kvm , it's now possible to deploy
CaaSP 4 in a libvirt environment. This is done with the terraform
libvirt provider and the terraform code from the SUSE/skuba project.

Note: The CaaSP4 deployment is still not integrated with the rest of
socok8s (this is true for deploying on OpenStack *and* libvirt).
But allowing to deploy CaaSP4 on libvirt is another step to allow
people to play with CaaSP4.