Don't try to remove a non-existent caaspv4 cluster
Attempting to run a teardown action on a default deployment fails due to
a missing terraform binary. Remove the caaspv4 cleanup from the teardown
action since a default deployment will deploy caaspv3. The caaspv3
cleanup can be replaced with the caaspv4 cleanup when we update the
setup_hosts action to deploy caaspv4 instead of caaspv3.