Delegate cert copy to the host with the cert
We generate the docker registry certificate on the deployer host, in the
event that the ansible runner host is not the deployer host, the copy
task will fail because there is no certificate on the ansible localhost.
Delegate the task to the first host in the host group we generated the
certificate on.