Remove references to inventory_hostname
Airship role is using inventory_hostname for hostnames.

This is a problem, as it represents the inventory
name, not the real node name. Kubernetes is using node name.

This fixes it by removing all the references in the airship side
of inventory_hostname to the ansible_hostname, which will be
gathered on the hosts by ansible, during the fact gathering step.