Use manual ssh-agent for caasp4 deploy
As done on upstream skuba[0] we need an ssh-agent
with a ssh key to be able to deploy the cluster and login
into it with skuba for cluster creation.

So adapt the script the caasp team uses upstream for skuba CI
testing which launches a ssh-agent instance and adds a known key
and inject that same key into the test instances via terraform.
We try to inject into the instances both a hardcoded key for jenkins
and the current user running the playbook key so this can work in
both CI and development machines

We modify the script to accept a parameter so we can use it with
several PRs at the same time, and to store the vars into a file so
we can source it on the necessary steps

[0] https://github.com/SUSE/skuba/blob/master/ci/tasks/setup-ssh.py