Report custom jobs to github
Use github-pr to report a combination of delivery mechanism +
caasp version to github in order to allow different jobs to run and
report back.

This allows us to run a combination of params for the same PR and
have easy forced checks for different paths, like airship+caasp3
and airship+caasp4, plus a variety of more combinations in the
future.

We could even have 2 jobs per PR in the future automatically,
but this is the first step to support that.

As a downside the normal context that reports automatically cannot
be disabled from jenkins as it would affect all other jobs that
use the github plugin, so instead we manage it ourselves and set it
to success at the start of the job as we create the other contexts,
so if this gets merged we should disable that mandatory check and enable
the airship+caasp3 instead