Rework airship-configure-worker for caasp4
Rework the role airship-configure-worker to split the tasks into 3
distinct parts (subvolumes, apparmor and enrolling nodes)
Both subvolumes and apparmor has to be run for caasp3 nodes only
and it needs to be run on all nodes.
Meanwhile enrolling the nodes can be done from one location.

This splits and calls the different tasks to run in the proper nodes.

This changes the node that runs the enrolling as in caasp4 we no
longer have the kubectl file configured in the caasp workers so we
need to use the deployer to run the enrolling.

This also makes it much simple and easy to rework and clean the
enrolling tasks to loop with the nodes list