Simplify CaaSP4 setup
There are dependencies on the localhost for deployment of CaaSP:
- We need a certain version of terraform into the localhost
  (which is not latest)
- We need to have certain repositories for skuba on localhost, some
  which aren't available outside SLE 15.

This simplifies everything by relying on a IBS built image, which
contains all the dependencies as caasp moves along. The only
dependency remains on podman, which runs the container 'containing'
the software necessary for caasp4.

Ansible dependency was removed and replaced by a simple shell
script.