Locally build site and transfer it to deployer
Ansible template is a bit slower, especially when doing local to
remote as it needs to check if the files exists, then check if
its different from the computed one and then template+transfer it
to the remote node.

In my local tests building the site from the templates could take from
30 minutes on a good connection to almost an hour in a connection with
a high latency (~1s).

Instead, lets just build the whole site in localhost under the workspace
and then transfer it on one go to the remote machine.

Currently I clean up the local copy of the site to keep everything
clean, but potentially we could leave those files up for ease of
checking what the site looks like without having to log into the
deployer. But this can come later, currently it just cleans up
everything at the end so we dont polute the workspace more than its
needed.

This results in an astonishing speed increase, taking only about 40s
in my high latency scenario (~1s). No downsides that I can see in this
approach