Cleanup CI workspace
Currently we are using viertualenv for no reason, as all packages
are available on the jenkins node already so we are generating
about 300Mb of virtualenv per run, which grows over time and
fills up the worker space.
On the same hand, we are generating about 200Mb of .tox dir for
the linting and docs, which then its never needed afterwards.

So lets just drop the virtualenv and add a new step to clean up
the .tox directory so we have a cleaner workspace afterwards